pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.9.0",
    "certifi>=2025.6.15",
    "cryptg>=0.5.0.post0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.27,<0.29",
    "idna>=3.10",
    "python-telegram-bot==22.2",
    "sniffio>=1.3.1",
    "telethon>=1.40.0",
    "typing-extensions>=4.14.0",
]
